requirements.txt
openai
playwright
weasyprint
markdown
python-dotenv
